+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file. This change
 ### Added
 - Convenience scripts to invoke Blutwurst in both bash and Windows Cmd.
 - Wrote a basic user manual in `doc/intro.md`.
-- Added city and email value generators.
+- Added phone, city and email value generators.
 
 ### Changed
 - Change string generation to be lorem ipsum based.